# Use same accumulation precision in gemv as gemm