--- a/mlx/backend/metal/kernels/gemv.metal
+++ b/mlx/backend/metal/kernels/gemv.metal
@@ -61,28 +61,32 @@ struct GEMVKernel {
   MLX_MTL_CONST short tgp_mem_size = BN > 1 ? BN*(blockM + TM) : 0;
   MLX_MTL_CONST bool needs_tgp_reduction = BN > 1;
 
+  template <typename U = T>
   static METAL_FUNC void
-  load_unsafe(const device T* src, thread T dst[TN], const int src_offset = 0) {
+  load_unsafe(const device T* src, thread U dst[TN], const int src_offset = 0) {
     MLX_MTL_PRAGMA_UNROLL
     for (int tn = 0; tn < TN; tn++) {
-      dst[tn] = src[src_offset + tn];
+      dst[tn] = static_cast<U>(src[src_offset + tn]);
     }
   }
 
+  template <typename U = T>
   static METAL_FUNC void load_safe(
       const device T* src,
-      thread T dst[TN],
+      thread U dst[TN],
       const int src_offset = 0,
       const int src_size = TN) {
     if (src_offset + TN <= src_size) {
       MLX_MTL_PRAGMA_UNROLL
       for (int tn = 0; tn < TN; tn++) {
-        dst[tn] = src[src_offset + tn];
+        dst[tn] = static_cast<U>(src[src_offset + tn]);
       }
     } else { // Edgecase
       MLX_MTL_PRAGMA_UNROLL
       for (int tn = 0; tn < TN; tn++) {
-        dst[tn] = src_offset + tn < src_size ? src[src_offset + tn] : 0;
+        dst[tn] = src_offset + tn < src_size
+            ? static_cast<U>(src[src_offset + tn])
+            : 0;
       }
     }
   }
@@ -109,7 +113,7 @@ struct GEMVKernel {
     // Thread local accumulation results
     thread AccT result[TM] = {0};
     thread T inter[TN];
-    thread T v_coeff[TN];
+    thread AccT v_coeff[TN];
 
     const int thrM = SN != 32 ? simd_lid / SN : 0;
     const int thrN = SN != 32 ? simd_lid % SN : int(simd_lid);
@@ -143,7 +147,7 @@ struct GEMVKernel {
 
     // Loop over in_vec in blocks of blockN
     for (int i = 0; i < n_iter; ++i) {
-      load_unsafe(in_vec, v_coeff, bn);
+      load_unsafe<AccT>(in_vec, v_coeff, bn);
 
       // Per thread work loop
       int mat_offset = 0;
@@ -155,7 +159,7 @@ struct GEMVKernel {
         // Accumulate results
         MLX_MTL_PRAGMA_UNROLL
         for (int tn = 0; tn < TN; tn++) {
-          result[tm] += static_cast<AccT>(inter[tn]) * v_coeff[tn];
+          result[tm] += inter[tn] * v_coeff[tn];
         }
 
         mat_offset += matrix_ld;
@@ -165,7 +169,7 @@ struct GEMVKernel {
     }
 
     if (leftover > 0) {
-      load_safe(in_vec, v_coeff, bn, in_size);
+      load_safe<AccT>(in_vec, v_coeff, bn, in_size);
 
       // Per thread work loop
       MLX_MTL_PRAGMA_UNROLL
@@ -176,7 +180,7 @@ struct GEMVKernel {
         // Accumulate results
         MLX_MTL_PRAGMA_UNROLL
         for (int tn = 0; tn < TN; tn++) {
-          result[tm] += static_cast<AccT>(inter[tn]) * v_coeff[tn];
+          result[tm] += inter[tn] * v_coeff[tn];
         }
       }
     }
@@ -296,8 +300,7 @@ struct GEMVTKernel {
     // Thread local accumulation results
     AccT result[TN] = {0};
     T inter[TN];
-    T v_coeff[TM];
-
+    AccT v_coeff[TM];
     const int thrM = SN != 32 ? simd_lid / SN : 0;
     const int thrN = SN != 32 ? simd_lid % SN : int(simd_lid);
 
@@ -333,16 +336,17 @@ struct GEMVTKernel {
 
         MLX_MTL_PRAGMA_UNROLL
         for (int tm = 0; tm < TM; tm++) {
-          v_coeff[tm] = in_vec[bm + tm];
+          v_coeff[tm] = static_cast<AccT>(in_vec[bm + tm]);
         }
 
         MLX_MTL_PRAGMA_UNROLL
         for (int tm = 0; tm < TM; tm++) {
+          auto vc = float(v_coeff[tm]);
           for (int tn = 0; tn < TN; tn++) {
             inter[tn] = mat[(bm + tm) * marix_ld + out_col + tn];
           }
           for (int tn = 0; tn < TN; tn++) {
-            result[tn] += static_cast<AccT>(v_coeff[tm]) * inter[tn];
+            result[tn] += vc * inter[tn];
           }
         }
 
@@ -351,7 +355,7 @@ struct GEMVTKernel {
 
       if (leftover > 0) {
         for (int tm = 0; tm < TM && bm + tm < in_vec_size; tm++) {
-          v_coeff[tm] = in_vec[bm + tm];
+          v_coeff[tm] = static_cast<AccT>(in_vec[bm + tm]);
 
           MLX_MTL_PRAGMA_UNROLL
           for (int tn = 0; tn < TN; tn++) {
@@ -360,7 +364,7 @@ struct GEMVTKernel {
 
           MLX_MTL_PRAGMA_UNROLL
           for (int tn = 0; tn < TN; tn++) {
-            result[tn] += static_cast<AccT>(v_coeff[tm]) * inter[tn];
+            result[tn] += v_coeff[tm] * inter[tn];
           }
         }
       }