This changes the gemv to accumulate in a higher precision which matches the behavior of gemm.

I think this is the right thing to do in general and will close #1958 

The benchmarks are confusing and need some investigation though.

For BF16 it's noticeably slower but for fp16 the speed is unchanged 🤔 . In both cases the math should be easily hidden by loading time so the bf16 result doesn't make sense.