@barronalex 2025-03-14
This seems very reasonable to me but probably worth @jagrit06 taking a look before we merge.
BF16 is quite a few more ops than fp16 if it's emulated right?