@awni 2025-03-14
I found a way to do this that doesn't result in much slow down for the important case `x @ W.T` and only slight slowdown for bf16 for the less important case (`x @ W`).

Type | Transposed | Time Pre | Time Post |
---- | ---- | ----- | -----
float16 | True |  2.772 (ms)| 2.778 (ms)
float16 | False | 8.019 (ms)| 8.025 (ms)
bfloat16 | True | 2.859 (ms)| 2.829 (ms)
bfloat16 | False |9.545 (ms)| 10.093 (ms)
float32 | True |  5.022 (ms)| 5.029 (ms)
float32 | False | 12.393 (ms)| 12.364 (ms)