@angeloskath 2025-03-14
I hesitated writing this yesterday (cause I only skimmed through the changes) but it makes sense cause we only do the casting once instead of every time we do a multiplication. On the other hand we need twice the registers so it could have a big impact on smaller machines. Let me run a quick benchmark on my M2 air.