@angeloskath 2025-03-14
I am running a square benchmark `1x4096 @ 4096x4096` but there is no difference on my laptop so I think it is good to go.

Before
```
Timing float32	x @ w	 ... 7.65121 msec
Timing float32	x @ w.T	 ... 7.88494 msec
Timing float16	x @ w	 ... 3.93161 msec
Timing float16	x @ w.T	 ... 3.98962 msec
Timing bfloat16	x @ w	 ... 3.97440 msec
Timing bfloat16	x @ w.T	 ... 3.97418 msec
```

After
```
Timing float32	x @ w	 ... 7.58956 msec
Timing float32	x @ w.T	 ... 7.77694 msec
Timing float16	x @ w	 ... 3.98002 msec
Timing float16	x @ w.T	 ... 3.93565 msec
Timing bfloat16	x @ w	 ... 3.95123 msec
Timing bfloat16	x @ w.T	 ... 4.01029 msec
```